refactor: move 4C related stuff to queens_interfaces/driver
